requirements.txt
openai>=1.35.7
anthropic>=0.34.2
mistralai>=0.4.1
python-dotenv>=1.0.1
requests>=2.32.3